--- a/benchmarks/bench-compare.js
+++ b/benchmarks/bench-compare.js
@@ -4,43 +4,13 @@ const Benchmark = require('benchmark')
 const SemVer = require('../classes/semver')
 const suite = new Benchmark.Suite()
 
-const versions = ['1.0.3', '2.2.2', '2.3.0']
-const versionToCompare = '1.0.2'
-const option1 = { includePrelease: true }
-const option2 = { includePrelease: true, loose: true }
-const option3 = { includePrelease: true, loose: true, rtl: true }
+const comparisons = require('../test/fixtures/comparisons')
 
-for (const version of versions) {
-  suite.add(`compare ${version} to ${versionToCompare}`, function () {
-    const semver = new SemVer(version)
-    semver.compare(versionToCompare)
+for (const [v0, v1] of comparisons) {
+  suite.add(`compare ${v0} to ${v1}`, function () {
+    const semver = new SemVer(v0)
+    semver.compare(v1)
   })
-}
-
-for (const version of versions) {
-  suite.add(
-    `compare ${version} to ${versionToCompare} with option (${JSON.stringify(option1)})`,
-    function () {
-      const semver = new SemVer(version, option1)
-      semver.compare(versionToCompare)
-    })
-}
-
-for (const version of versions) {
-  suite.add(`compare ${version} to ${versionToCompare} with option (${JSON.stringify(option2)})`,
-    function () {
-      const semver = new SemVer(version, option2)
-      semver.compare(versionToCompare)
-    })
-}
-
-for (const version of versions) {
-  suite.add(
-    `compare ${version} to ${versionToCompare} with option (${JSON.stringify(option3)})`,
-    function () {
-      const semver = new SemVer(version, option3)
-      semver.compare(versionToCompare)
-    })
 }
 
 suite