@wraithgar 2025-10-04
I think this is worth digging into just a little bit.

Under what situations is the identifier parseable as an integer but not cast as one?  This internal function is only ever used in the semver class itself.

 - Semver always casts `this.major`, `this.minor`, and `this.patch` to numbers.
 - It also casts any numbers in the `this.prerelease` array to numbers.
 - It also iterates through each entry in `this.prerelease` individually when passing to `compareIdentifiers`

I think what we have here is a classic pitfall of unit testing: by testing the internals file directly we have disregarded the ways in which it can actually be called.  Unless you start manually setting entries in the `this.prerelease` array, and setting them as strings representing a number, we don't actually ever need to cast these to numbers.

Unfortunately because of the coverage map.js file we can't do away with this test files itself.  What we can do is write tests in that file that don't require the internals, but get to the code from the actual Semver object.

TLDR: let's stop casting these things to numbers at all.  We can remove

```javascript
const numeric = /^[0-9]+$/
```

and

```javascript
  const anum = numeric.test(a)
  const bnum = numeric.test(b)

  if (anum && bnum) {
    a = +a
    b = +b
  }
```

What do you think?