@wraithgar 2025-10-04
`bench-compare.js` needs to have pre-release identifiers added to `versions`