@wraithgar 2025-10-05
Yeah `compareIdentifier` only makes sense when parsing prerelease components.  For major/minor/patch it's just standard number comparison. We can do that inline.

Check it out: https://github.com/npm/node-semver/commit/03315c7f543e62c33f75817d7e3a6589e78b1b12