@wraithgar 2025-10-06
> Change:
> 
> ```diff
> diff --git a/classes/semver.js b/classes/semver.js
> index 92254be..9d18f21 100644
> --- a/classes/semver.js
> +++ b/classes/semver.js
> @@ -64,10 +64,10 @@ class SemVer {
>        this.prerelease = []
>      } else {
>        this.prerelease = m[4].split('.').map((id) => {
> -        if (/^[0-9]+$/.test(id)) {
> -          const num = +id
> -          if (num >= 0 && num < MAX_SAFE_INTEGER) {
> -            return num
> +        if (id !== '') {
> +          const idNum = +id
> +          if (idNum >= 0 && idNum < MAX_SAFE_INTEGER) {
> +            return idNum
>            }
>          }
>          return id
> ```

Fun challenge: try to get `id` to be `''` in this map.