@wraithgar 2025-10-07
semver.compare is, on average, 36% faster now.  41% faster if there are no prereleases.

<img width="870" height="684" alt="Spreadsheet of performance benchmark before and after this version" src="https://github.com/user-attachments/assets/11480b20-ecb9-4940-b1c8-b161bf41ce11" />